@@ -3,7 +3,7 @@ library random_data;
 import 'common.dart';
 
 List<Offering> generateOfferings(int count) =>
-  new List.generate(count, generateOffering);
+    new List.generate(count, generateOffering);
 
 Offering generateOffering(int seed) {
   final res = new Offering();
@@ -22,24 +22,32 @@ Offering generateOffering(int seed) {
 }
 
 Company generateCompany(int seed) {
-  return new Company()
-    ..name = generateName(seed);
+  return new Company()..name = generateName(seed);
 }
 
 Opportunity generateOpportunity(int seed) {
-  return new Opportunity()
-    ..name = generateName(seed);
+  return new Opportunity()..name = generateName(seed);
 }
 
 Account generateAccount(int seed) {
-  return new Account()
-    ..accountId = seed;
+  return new Account()..accountId = seed;
 }
 
 String generateName(int seed) {
   const names = const [
-    'Foo', 'Bar', 'Baz', 'Qux', 'Quux', 'Garply', 'Waldo', 'Fred', 'Plugh',
-      'Xyzzy', 'Thud', 'Cruft', 'Stuff'
+    'Foo',
+    'Bar',
+    'Baz',
+    'Qux',
+    'Quux',
+    'Garply',
+    'Waldo',
+    'Fred',
+    'Plugh',
+    'Xyzzy',
+    'Thud',
+    'Cruft',
+    'Stuff'
   ];
   return names[seed % names.length];
 }
@@ -49,13 +57,13 @@ DateTime randomDate(int seed, {DateTime minDate}) {
     minDate = new DateTime.now();
   }
 
-  const offsets = const[0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
+  const offsets = const [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
   return minDate.add(new Duration(days: offsets[seed % offsets.length]));
 }
 
 String randomString(int seed) {
   return new String.fromCharCodes(new List.generate(
-    const[5, 7, 9, 11, 13][seed % 5],
-    (i) => 'a'.codeUnitAt(0) + const[0, 1, 2, 3, 4, 5, 6, 7, 8][seed % 9] + i
-  ));
+      const [5, 7, 9, 11, 13][seed % 5],
+      (i) =>
+          'a'.codeUnitAt(0) + const [0, 1, 2, 3, 4, 5, 6, 7, 8][seed % 9] + i));
 }
