@@ -1,2 +1,3 @@
 export './common.dart';
-export './src/webdriver/async_webdriver_adapter.dart' show AsyncWebDriverAdapter;
+export './src/webdriver/async_webdriver_adapter.dart'
+    show AsyncWebDriverAdapter;
