@@ -1,7 +1,7 @@
 library benchpress.src.webdriver.async_webdriver_adapter_dart;
 
+import 'dart:async';
 import 'package:webdriver/webdriver.dart' show WebDriver, LogEntry;
-import 'package:angular2/src/facade/async.dart' show Future;
 import '../web_driver_adapter.dart' show WebDriverAdapter;
 
 class AsyncWebDriverAdapter extends WebDriverAdapter {
@@ -13,7 +13,11 @@ class AsyncWebDriverAdapter extends WebDriverAdapter {
   }
 
   Future executeScript(String script) {
-    return _driver.execute(script, const[]);
+    return _driver.execute(script, const []);
+  }
+
+  Future executeAsyncScript(String script) {
+    return _driver.executeAsync(script, const []);
   }
 
   Future<Map> capabilities() {
@@ -21,12 +25,11 @@ class AsyncWebDriverAdapter extends WebDriverAdapter {
   }
 
   Future<List<Map>> logs(String type) {
-    return _driver.logs.get(type)
-      .map((LogEntry entry) => {
-        'message': entry.message
-      })
-      .fold(<Map>[], (log, Map entry) {
-        return log..add(entry);
-      });
+    return _driver.logs
+        .get(type)
+        .map((LogEntry entry) => {'message': entry.message})
+        .fold(<Map>[], (log, Map entry) {
+      return log..add(entry);
+    });
   }
 }
