@@ -1,7 +1,5 @@
-// Use "register" extension from systemjs.
-// That's what Traceur outputs: `System.register()`.
-register(System);
-cjs(System);
+// Tun on full stack traces in errors to help debugging
+Error.stackTraceLimit=Infinity;
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 100;
 
@@ -9,21 +7,80 @@ jasmine.DEFAULT_TIMEOUT_INTERVAL = 100;
 // we will call `__karma__.start()` later, once all the specs are loaded.
 __karma__.loaded = function() {};
 
-System.baseURL = '/base/';
+System.config({
+  baseURL: '/base',
+  defaultJSExtensions: true,
+  map: {
+    'benchpress/*': 'dist/js/dev/es5/benchpress/*.js',
+    '@angular': 'dist/all/@angular',
+    'rxjs': 'node_modules/rxjs',
+    'parse5/index': 'dist/all/empty.js',
+    '@angular/platform-server/src/parse5_adapter': 'dist/all/empty.js',
+    'angular2/*': 'dist/all/angular2/*.js',
+    'angular2/src/alt_router/router_testing_providers': 'dist/all/angular2/src/alt_router/router_testing_providers.js'
+  },
+  packages: {
+    '@angular/core': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/compiler': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/common': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/forms': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    // remove after all tests imports are fixed
+    '@angular/facade': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/router-deprecated': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/http': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/upgrade': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/platform-browser': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/platform-browser-dynamic': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    },
+    '@angular/platform-server': {
+      main: 'index.js',
+      defaultExtension: 'js'
+    }
+  }
+});
 
-// So that we can import packages like `core/foo`, instead of `core/src/foo`.
-System.paths = {
-  '*': './*.js',
-  'benchpress/*': 'dist/js/dev/es5/benchpress/*.js',
-  'angular2/*': 'dist/js/dev/es5/angular2/*.js',
-  'rtts_assert/*': 'dist/js/dev/es5/rtts_assert/*.js',
-  'rx': 'node_modules/rx/dist/rx.js'
-};
 
-// Import all the specs, execute their `main()` method and kick off Karma (Jasmine).
-System.import('angular2/src/dom/browser_adapter').then(function(browser_adapter) {
-  browser_adapter.BrowserDomAdapter.makeCurrent();
-}).then(function() {
+// Set up the test injector, then import all the specs, execute their `main()`
+// method and kick off Karma (Jasmine).
+System.import('@angular/core/testing')
+  .then(function(coreTesting){
+    return System.import('@angular/platform-browser-dynamic/testing')
+      .then(function(browserTesting) {
+         coreTesting.initTestEnvironment(
+           browserTesting.BrowserDynamicTestModule,
+           browserTesting.browserDynamicTestPlatform());
+      });
+  })
+.then(function() {
   return Promise.all(
     Object.keys(window.__karma__.files) // All files served by Karma.
     .filter(onlySpecFiles)
@@ -41,8 +98,7 @@ System.import('angular2/src/dom/browser_adapter').then(function(browser_adapter)
 .then(function() {
   __karma__.start();
 }, function(error) {
-  console.error(error.stack || error);
-  __karma__.start();
+  __karma__.error(error.stack || error);
 });
 
 
