@@ -1,6 +1,6 @@
-/// <reference path="../typings/es6-promise/es6-promise.d.ts" />
-
-
-declare class Writer { write(readTree: (tree) => Promise<string>, destDir: string): Promise<any>; }
-
-export = Writer;
+declare module 'broccoli-writer' {
+  class Writer {
+    write(readTree: (tree: BroccoliTree) => Promise<string>, destDir: string): Promise<any>;
+  }
+  export = Writer;
+}
