@@ -1,6 +1,3 @@
-/// <reference path="../typings/es6-promise/es6-promise.d.ts" />
-
-
 interface BroccoliTree {
   /**
    * Contains the fs path for the input tree when the plugin takes only one input tree.
@@ -17,7 +14,8 @@ interface BroccoliTree {
    * For plugins that take only one input tree, it might be more convenient to use the `inputPath`
    *property instead.
    *
-   * This property is set just before the first rebuild and doesn't change afterwards.
+   * This property is set just before the first rebuild and doesn't change afterwards, unless
+   * plugins themselves change it.
    *
    * If the inputPath is outside of broccoli's temp directory, then it's lifetime is not managed by
    *the builder.
@@ -31,7 +29,8 @@ interface BroccoliTree {
   /**
    * Contains the fs paths for the output trees.
    *
-   * This property is set just before the first rebuild and doesn't change afterwards.
+   * This property is set just before the first rebuild and doesn't change afterwards, unless the
+   * plugins themselves change it.
    *
    * The underlying directory is also created by the builder just before the first rebuild.
    * This directory is destroyed and recreated upon each rebuild.
@@ -52,19 +51,25 @@ interface BroccoliTree {
   inputTrees?: BroccoliTree[];
 
   /**
+   * Trees which implement the rebuild api are wrapped automatically for api compat,
+   * and `newStyleTree` keeps a reference to the original unwrapped tree.
+   */
+  newStyleTree?: BroccoliTree;
+
+  /**
    * Description or name of the plugin used for reporting.
    *
    * If missing `tree.constructor.name` is usually used instead.
    */
   description?: string;
 
-  rebuild(): (Promise<any>| void);
+  rebuild(): (Promise<any>|void);
   cleanup(): void;
 }
 
 
 interface OldBroccoliTree {
-  read ? (readTree: (tree: BroccoliTree) => Promise<string>) : (Promise<string>| string);
+  read?(readTree: (tree: BroccoliTree) => Promise<string>): (Promise<string>|string);
 }
 
 
